# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OwlMail HTTP API",
    "version": "0.6.0",
    "summary": "Versioned OwlMail mailbox, relay, settings, and operational API",
    "description": "The contract for OwlMail's native `/api/v1` interface. Unversioned MailDev-style compatibility routes are intentionally excluded because their protocol behavior is not an exact MailDev contract. Configure `-base-pathname` (or `OWLMAIL_BASE_PATHNAME`) to place the server and this contract below a reverse-proxy prefix. The served documents rewrite the relative server URL to include that prefix. HTTP Basic authentication is optional at deployment time; when enabled it protects every operation except liveness and readiness. The unversioned `/healthz` and `/readyz` probes are also public but remain outside this versioned contract."
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Root deployment; the served contract includes the configured base pathname"
    }
  ],
  "security": [
    {
      "basicAuth": []
    },
    {}
  ],
...
  "paths": {
    "/openapi.json": {
      "get": {
        "tags": [
          "Contract"
        ],
        "summary": "Get the OpenAPI contract as JSON",
        "operationId": "getOpenAPIJSON",
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 JSON document with the effective base path in `servers`.",
            "content": {
              "application/vnd.oai.openapi+json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "The embedded contract could not be rendered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.yaml": {
      "get": {
        "tags": [
          "Contract"
        ],
        "summary": "Get the OpenAPI contract as YAML",
        "operationId": "getOpenAPIYAML",
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 YAML document with the effective base path in `servers`.",
            "content": {
              "application/vnd.oai.openapi": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "The embedded contract could not be rendered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/emails": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List full emails",
        "operationId": "listEmails",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Search"
          },
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/DateFrom"
          },
          {
            "$ref": "#/components/parameters/DateTo"
          },
          {
            "$ref": "#/components/parameters/Read"
          },
          {
            "$ref": "#/components/parameters/SortBy"
          },
          {
            "$ref": "#/components/parameters/SortOrder"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailList"
                },
                "example": {
                  "total": 1,
                  "limit": 20,
                  "offset": 0,
                  "emails": [
                    {
                      "id": "aB3dEfGh",
                      "time": "2026-09-02T12:00:00Z",
                      "read": false,
                      "subject": "Welcome",
                      "from": [
                        {
                          "Name": "Sender",
                          "Address": "sender@example.com"
                        }
                      ],
                      "to": [
                        {
                          "Name": "",
                          "Address": "recipient@example.com"
                        }
                      ],
                      "cc": [],
                      "bcc": [],
                      "calculatedBcc": [],
                      "text": "Hello",
                      "html": "<p>Hello</p>",
                      "attachments": [],
                      "envelope": {
                        "from": "sender@example.com",
                        "to": [
                          "recipient@example.com"
                        ],
                        "cc": [],
                        "bcc": [],
                        "calculatedBcc": [],
                        "host": "localhost",
                        "remoteAddress": "127.0.0.1:2525"
                      },
                      "source": "",
                      "size": 128,
                      "sizeHuman": "128 B",
                      "headers": {}
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete all emails",
        "operationId": "deleteAllEmails",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                },
                "example": {
                  "code": "ALL_EMAILS_DELETED",
                  "message": "All emails deleted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Mailbox deletion failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/emails/stats": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get mailbox statistics",
        "operationId": "getEmailStats",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStats"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/emails/preview": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List lightweight email previews",
        "operationId": "listEmailPreviews",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Search"
          },
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/DateFrom"
          },
          {
            "$ref": "#/components/parameters/DateTo"
          },
          {
            "$ref": "#/components/parameters/Read"
          },
          {
            "$ref": "#/components/parameters/SortBy"
          },
          {
            "$ref": "#/components/parameters/SortOrder"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPreviewList"
                },
                "example": {
                  "total": 1,
                  "limit": 50,
                  "offset": 0,
                  "previews": [
                    {
                      "id": "aB3dEfGh",
                      "time": "2026-09-02T12:00:00Z",
                      "read": false,
                      "subject": "Welcome",
                      "from": "sender@example.com",
                      "to": [
                        "recipient@example.com"
                      ],
                      "size": 128,
                      "sizeHuman": "128 B",
                      "hasAttachment": false,
                      "preview": "Hello"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/emails/export": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Export matching EML files as ZIP",
        "description": "Exports at most 1000 readable messages and 256 MiB of raw source.",
        "operationId": "exportEmails",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExportIDs"
          },
          {
            "$ref": "#/components/parameters/Search"
          },
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/DateFrom"
          },
          {
            "$ref": "#/components/parameters/DateTo"
          },
          {
            "$ref": "#/components/parameters/Read"
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP archive of EML files.",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "description": "The export exceeds the message or byte limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/emails/read": {
      "patch": {
        "tags": [
          "Emails"
        ],
        "summary": "Mark all emails as read",
        "operationId": "readAllEmails",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                },
                "example": {
                  "code": "ALL_EMAILS_MARKED_READ",
                  "message": "All emails marked as read",
                  "data": {
                    "count": 2
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "The mailbox update failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/emails/batch": {
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete a batch of emails",
        "operationId": "batchDeleteEmails",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              },
              "example": {
                "ids": [
                  "aB3dEfGh",
                  "missing"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                },
                "example": {
                  "code": "BATCH_DELETE_COMPLETED",
                  "message": "Batch delete completed",
                  "success": 1,
                  "failed": 1,
                  "failedIDs": [
                    "missing"
                  ],
                  "total": 2
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "description": "More than 1000 email IDs were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/emails/batch/read": {
      "patch": {
        "tags": [
          "Emails"
        ],
        "summary": "Mark a batch of emails as read",
        "operationId": "batchReadEmails",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              },
              "example": {
                "ids": [
                  "aB3dEfGh"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResult"
                },
                "example": {
                  "code": "BATCH_READ_COMPLETED",
                  "message": "Batch read completed",
                  "success": 1,
                  "failed": 0,
                  "failedIDs": [],
                  "total": 1
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "description": "More than 1000 email IDs were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Deployment-conditional HTTP Basic Auth. When server credentials are configured, callers must supply matching credentials; rejected requests receive a plain-text 401 with `WWW-Authenticate: Basic realm=\"OwlMail\"`."
      }
    },
    "parameters": {
...
      "Search": {
        "name": "q",
        "in": "query",
        "description": "Case-insensitive substring search in subject, text, and HTML.",
        "schema": {
          "type": "string"
        },
        "example": "welcome"
      },
...
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Basic Auth is enabled and credentials are missing or invalid. The body is plain text, not the JSON API error structure.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "Present as `Basic realm=\"OwlMail\"` when the Authorization header is missing."
          }
        },
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            },
            "example": "Unauthorized"
          }
        }
      },
...
    }
  }
}